requirements.txt
dj-database-url==0.5.0
Django==3.1.3
django-allauth==0.45.0
django-cors-headers==3.8.0
django-heroku==0.3.1
djangorestframework==3.12.4
gunicorn==20.0.4
psycopg2==2.8.6
pytz==2020.4
requests==2.21.0
requests-oauthlib==1.3.0
whitenoise==5.2.0
